# openapi.yaml
openapi: 3.0.3
info:
  title: PulseTrade API
  version: 1.0.0
  description: >
    Refreshed public API surface. Auth is session cookie (Flask-Login) or
    JWT Bearer depending on endpoint; admin routes require admin session.
    Endpoints marked demo return synthetic data when DEMO=1 and no
    exchange keys or local JSON dumps exist.
servers:
  - url: /
paths:
...
  /getPrice/:
    get:
      summary: Price + indicator snapshot for a symbol (demo fallback available)
      parameters:
        - { name: symbol, in: query, schema: { type: string } }
        - { name: interval, in: query, schema: { type: string } }
        - { name: exchange, in: query, schema: { type: string } }
      responses: { '200': { description: snapshot } }
  /openOrders/:
    get:
      summary: Cached open orders (demo fallback available)
      responses: { '200': { description: ok }, '404': { description: no data } }
  /lastTrades/:
    get:
      summary: Cached recent trades (demo fallback available)
      responses: { '200': { description: ok }, '404': { description: no data } }
...